Dedicated package for extension CRD deletion validator webhook
