+++ pkg/seedadmission/seedadmission_suite_test.go
@@ -16,8 +16,10 @@ package seedadmission_test
 
 import (
 	"context"
-	"net/http"
+	"fmt"
+	"path/filepath"
 	"testing"
+	"time"
 
 	"github.com/go-logr/logr"
 	. "github.com/onsi/ginkgo"
@@ -30,10 +32,13 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	apiextensionsv1beta1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/rest"
 	"k8s.io/utils/pointer"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	logzap "sigs.k8s.io/controller-runtime/pkg/log/zap"
@@ -44,13 +49,17 @@ import (
 	"github.com/gardener/gardener/cmd/utils"
 	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
+	"github.com/gardener/gardener/pkg/client/kubernetes"
+	"github.com/gardener/gardener/pkg/operation/common"
 	"github.com/gardener/gardener/pkg/seedadmission"
+	"github.com/gardener/gardener/pkg/seedadmission/webhooks/admission/extensioncrds"
+	"github.com/gardener/gardener/pkg/utils/test"
 	"github.com/gardener/gardener/test/framework"
 )
 
-func TestSeedadmission(t *testing.T) {
+func TestSeedAdmissionController(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Seed Admission Suite")
+	RunSpecs(t, "SeedAdmissionController Suite")
 }
 
 var (
@@ -92,7 +101,7 @@ var _ = BeforeSuite(func() {
 
 	By("setting up webhook server")
 	server := mgr.GetWebhookServer()
-	server.Register(seedadmission.ExtensionDeletionProtectionWebhookPath, &webhook.Admission{Handler: &seedadmission.ExtensionDeletionProtection{}})
+	server.Register(extensioncrds.WebhookPath, &webhook.Admission{Handler: extensioncrds.New(logger)})
 	server.Register(seedadmission.GardenerShootControlPlaneSchedulerWebhookPath, &webhook.Admission{Handler: admission.HandlerFunc(seedadmission.DefaultShootControlPlanePodsSchedulerName)})
 
 	go func() {
@@ -132,7 +141,7 @@ func getValidatingWebhookConfig() *admissionregistrationv1beta1.ValidatingWebhoo
 			}},
 			ClientConfig: admissionregistrationv1beta1.WebhookClientConfig{
 				Service: &admissionregistrationv1beta1.ServiceReference{
-					Path: pointer.StringPtr(seedadmission.ExtensionDeletionProtectionWebhookPath),
+					Path: pointer.StringPtr(extensioncrds.WebhookPath),
 				},
 			},
 			AdmissionReviewVersions: []string{admissionv1beta1.SchemeGroupVersion.Version},
@@ -158,13 +167,181 @@ func getValidatingWebhookConfig() *admissionregistrationv1beta1.ValidatingWebhoo
 			}},
 			ClientConfig: admissionregistrationv1beta1.WebhookClientConfig{
 				Service: &admissionregistrationv1beta1.ServiceReference{
-					Path: pointer.StringPtr(seedadmission.ExtensionDeletionProtectionWebhookPath),
+					Path: pointer.StringPtr(extensioncrds.WebhookPath),
 				},
 			},
 			AdmissionReviewVersions: []string{admissionv1beta1.SchemeGroupVersion.Version},
 		}},
 	}
 }
+
+var _ = Describe("Integration Test", func() {
+	Describe("Extension CRDs Webhook Handler", func() {
+		var (
+			c         client.Client
+			namespace = "shoot--foo--bar"
+
+			crdObjects = []client.Object{
+				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "backupbuckets.extensions.gardener.cloud"}},
+				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "backupentries.extensions.gardener.cloud"}},
+				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "containerruntimes.extensions.gardener.cloud"}},
+				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "controlplanes.extensions.gardener.cloud"}},
+				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "extensions.extensions.gardener.cloud"}},
+				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "infrastructures.extensions.gardener.cloud"}},
+				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "networks.extensions.gardener.cloud"}},
+				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "operatingsystemconfigs.extensions.gardener.cloud"}},
+				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "workers.extensions.gardener.cloud"}},
+			}
+			extensionObjects = []client.Object{
+				&extensionsv1alpha1.BackupBucket{ObjectMeta: metav1.ObjectMeta{Name: "foo"}},
+				&extensionsv1alpha1.BackupEntry{ObjectMeta: metav1.ObjectMeta{Name: namespace}},
+				&extensionsv1alpha1.ContainerRuntime{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
+				&extensionsv1alpha1.ControlPlane{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
+				&extensionsv1alpha1.Extension{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
+				&extensionsv1alpha1.Infrastructure{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
+				&extensionsv1alpha1.Network{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
+				&extensionsv1alpha1.OperatingSystemConfig{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
+				&extensionsv1alpha1.Worker{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
+			}
+			podObject = &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{Name: "foo", Namespace: namespace},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{{
+						Name:  "foo",
+						Image: "foo:latest",
+					}},
+				},
+			}
+
+			deletionUnprotectedLabels    = map[string]string{common.GardenerDeletionProtected: "false"}
+			deletionConfirmedAnnotations = map[string]string{common.ConfirmationDeletion: "true"}
+		)
+
+		BeforeEach(func() {
+			c, err = client.New(restConfig, client.Options{Scheme: kubernetes.SeedScheme})
+			Expect(err).NotTo(HaveOccurred())
+
+			ns := &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}
+			if err := c.Create(ctx, ns); err != nil && !apierrors.IsAlreadyExists(err) {
+				Expect(err).NotTo(HaveOccurred())
+			}
+
+			By("applying CRDs")
+			applier, err := kubernetes.NewChartApplierForConfig(restConfig)
+			Expect(err).NotTo(HaveOccurred())
+			repoRoot := filepath.Join("..", "..")
+			Expect(applier.Apply(ctx, filepath.Join(repoRoot, "charts", "seed-bootstrap", "charts", "extensions"), "extensions", "")).To(Succeed())
+
+			Eventually(func() bool {
+				for _, object := range extensionObjects {
+					err := c.Get(ctx, client.ObjectKeyFromObject(object), object)
+					if meta.IsNoMatchError(err) {
+						return false
+					}
+				}
+				return true
+			}, 1*time.Second, 50*time.Millisecond).Should(BeTrue())
+		})
+
+		objectID := func(obj client.Object) string {
+			return fmt.Sprintf("%T/%s", obj, client.ObjectKeyFromObject(obj))
+		}
+
+		testDeletionUnconfirmed := func(ctx context.Context, obj client.Object) {
+			Eventually(func() string {
+				err := c.Delete(ctx, obj)
+				return string(apierrors.ReasonForError(err))
+			}, 1*time.Second, 50*time.Millisecond).Should(ContainSubstring("annotation to delete"), objectID(obj))
+		}
+
+		testDeletionConfirmed := func(ctx context.Context, obj client.Object) {
+			Eventually(func() error {
+				return c.Delete(ctx, obj)
+			}, 1*time.Second, 50*time.Millisecond).ShouldNot(HaveOccurred(), objectID(obj))
+			Eventually(func() bool {
+				err := c.Get(ctx, client.ObjectKeyFromObject(obj), obj)
+				return apierrors.IsNotFound(err) || meta.IsNoMatchError(err)
+			}, 1*time.Second, 50*time.Millisecond).Should(BeTrue(), objectID(obj))
+		}
+
+		Context("custom resource definitions", func() {
+			It("should admit the deletion because CRD has no protection label", func() {
+				for _, obj := range crdObjects {
+					// patch out default gardener.cloud/deletion-protected=true label
+					_, err := controllerutil.CreateOrPatch(ctx, c, obj, func() error {
+						obj.SetLabels(nil)
+						return nil
+					})
+					Expect(err).NotTo(HaveOccurred(), objectID(obj))
+					testDeletionConfirmed(ctx, obj)
+				}
+			})
+
+			It("should admit the deletion because CRD's protection label is not true", func() {
+				for _, obj := range crdObjects {
+					// override default gardener.cloud/deletion-protected=true label
+					_, err := controllerutil.CreateOrPatch(ctx, c, obj, func() error {
+						obj.SetLabels(deletionUnprotectedLabels)
+						return nil
+					})
+					Expect(err).NotTo(HaveOccurred(), objectID(obj))
+					testDeletionConfirmed(ctx, obj)
+				}
+			})
+
+			It("should prevent the deletion because CRD's protection label is true but deletion is not confirmed", func() {
+				// CRDs in seed-bootstrap chart should have gardener.cloud/deletion-protected=true label by default
+				for _, obj := range crdObjects {
+					testDeletionUnconfirmed(ctx, obj)
+				}
+			})
+
+			It("should admit the deletion because CRD's protection label is true and deletion is confirmed", func() {
+				// CRDs in seed-bootstrap chart should have gardener.cloud/deletion-protected=true label by default
+				for _, obj := range crdObjects {
+					_, err := controllerutil.CreateOrPatch(ctx, c, obj, func() error {
+						obj.SetAnnotations(deletionConfirmedAnnotations)
+						return nil
+					})
+					Expect(err).NotTo(HaveOccurred(), objectID(obj))
+					testDeletionConfirmed(ctx, obj)
+				}
+			})
+		})
+
+		Context("extension resources", func() {
+			BeforeEach(func() {
+				By("creating extension test objects")
+				_, err := test.EnsureTestResources(ctx, c, "testdata")
+				Expect(err).NotTo(HaveOccurred())
+			})
+
+			It("should prevent the deletion because deletion is not confirmed", func() {
+				for _, obj := range extensionObjects {
+					testDeletionUnconfirmed(ctx, obj)
+				}
+			})
+
+			It("should admit the deletion because deletion is confirmed", func() {
+				for _, obj := range extensionObjects {
+					_, err := controllerutil.CreateOrPatch(ctx, c, obj, func() error {
+						obj.SetAnnotations(deletionConfirmedAnnotations)
+						return nil
+					})
+					Expect(err).NotTo(HaveOccurred(), objectID(obj))
+					testDeletionConfirmed(ctx, obj)
+				}
+			})
+		})
+
+		Context("other resources", func() {
+			It("should not block deletion of other resources", func() {
+				Expect(c.Create(ctx, podObject)).To(Succeed())
+				testDeletionConfirmed(ctx, podObject)
+			})
+		})
+	})
+})
 
 func getMutatingWebhookConfig() *admissionregistrationv1beta1.MutatingWebhookConfiguration {
 	scope := admissionregistrationv1beta1.NamespacedScope
@@ -211,16 +388,4 @@ func expectAllowed(response admission.Response, reason gomegatypes.GomegaMatcher
 func expectPatched(response admission.Response, reason gomegatypes.GomegaMatcher, patches []jsonpatch.JsonPatchOperation, optionalDescription ...interface{}) {
 	expectAllowed(response, reason, optionalDescription...)
 	Expect(response.Patches).To(Equal(patches))
-}
-
-func expectDenied(response admission.Response, reason gomegatypes.GomegaMatcher, optionalDescription ...interface{}) {
-	Expect(response.Allowed).To(BeFalse(), optionalDescription...)
-	Expect(response.Result.Code).To(BeEquivalentTo(http.StatusForbidden), optionalDescription...)
-	Expect(string(response.Result.Reason)).To(reason, optionalDescription...)
-}
-
-func expectErrored(response admission.Response, code, err gomegatypes.GomegaMatcher, optionalDescription ...interface{}) {
-	Expect(response.Allowed).To(BeFalse(), optionalDescription...)
-	Expect(response.Result.Code).To(code, optionalDescription...)
-	Expect(response.Result.Message).To(err, optionalDescription...)
 }
