+++ pkg/seedadmission/webhooks/admission/extensioncrds/admission_test.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
+// Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package seedadmission_test
+package extensioncrds_test
 
 import (
 	"context"
@@ -20,23 +20,21 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
-	"path/filepath"
-	"time"
 
+	"github.com/go-logr/logr"
 	"github.com/golang/mock/gomock"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	gomegatypes "github.com/onsi/gomega/types"
 	admissionv1 "k8s.io/api/admission/v1"
-	corev1 "k8s.io/api/core/v1"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	apiextensionsv1beta1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+	logzap "sigs.k8s.io/controller-runtime/pkg/log/zap"
 	"sigs.k8s.io/controller-runtime/pkg/runtime/inject"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
@@ -45,14 +43,16 @@ import (
 	"github.com/gardener/gardener/pkg/client/kubernetes"
 	mockclient "github.com/gardener/gardener/pkg/mock/controller-runtime/client"
 	"github.com/gardener/gardener/pkg/operation/common"
-	. "github.com/gardener/gardener/pkg/seedadmission"
+	"github.com/gardener/gardener/pkg/seedadmission/webhooks/admission/extensioncrds"
 	kutil "github.com/gardener/gardener/pkg/utils/kubernetes"
-	"github.com/gardener/gardener/pkg/utils/test"
 )
 
-var _ = Describe("Extension Deletion Protection", func() {
+var _ = Describe("plugin", func() {
 	Describe("#ValidateExtensionDeletion", func() {
 		var (
+			ctx    = context.TODO()
+			logger logr.Logger
+
 			request admission.Request
 			decoder *admission.Decoder
 
@@ -63,6 +63,8 @@ var _ = Describe("Extension Deletion Protection", func() {
 		)
 
 		BeforeEach(func() {
+			logger = logzap.New(logzap.WriteTo(GinkgoWriter))
+
 			ctrl = gomock.NewController(GinkgoT())
 			c = mockclient.NewMockClient(ctrl)
 
@@ -73,8 +75,7 @@ var _ = Describe("Extension Deletion Protection", func() {
 			decoder, err = admission.NewDecoder(kubernetes.SeedScheme)
 			Expect(err).NotTo(HaveOccurred())
 
-			validator = &ExtensionDeletionProtection{}
-			Expect(inject.LoggerInto(logger, validator)).To(BeTrue())
+			validator = extensioncrds.New(logger)
 			Expect(inject.APIReaderInto(c, validator)).To(BeTrue())
 			Expect(admission.InjectDecoderInto(decoder, validator)).To(BeTrue())
 		})
@@ -575,170 +576,21 @@ var _ = Describe("Extension Deletion Protection", func() {
 			})
 		})
 	})
-
-	Describe("Integration Test", func() {
-		var (
-			c         client.Client
-			namespace = "shoot--foo--bar"
-
-			crdObjects = []client.Object{
-				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "backupbuckets.extensions.gardener.cloud"}},
-				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "backupentries.extensions.gardener.cloud"}},
-				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "containerruntimes.extensions.gardener.cloud"}},
-				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "controlplanes.extensions.gardener.cloud"}},
-				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "extensions.extensions.gardener.cloud"}},
-				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "infrastructures.extensions.gardener.cloud"}},
-				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "networks.extensions.gardener.cloud"}},
-				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "operatingsystemconfigs.extensions.gardener.cloud"}},
-				&apiextensionsv1.CustomResourceDefinition{ObjectMeta: metav1.ObjectMeta{Name: "workers.extensions.gardener.cloud"}},
-			}
-			extensionObjects = []client.Object{
-				&extensionsv1alpha1.BackupBucket{ObjectMeta: metav1.ObjectMeta{Name: "foo"}},
-				&extensionsv1alpha1.BackupEntry{ObjectMeta: metav1.ObjectMeta{Name: namespace}},
-				&extensionsv1alpha1.ContainerRuntime{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
-				&extensionsv1alpha1.ControlPlane{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
-				&extensionsv1alpha1.Extension{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
-				&extensionsv1alpha1.Infrastructure{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
-				&extensionsv1alpha1.Network{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
-				&extensionsv1alpha1.OperatingSystemConfig{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
-				&extensionsv1alpha1.Worker{ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: "foo"}},
-			}
-			podObject = &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{Name: "foo", Namespace: namespace},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{{
-						Name:  "foo",
-						Image: "foo:latest",
-					}},
-				},
-			}
-
-			deletionUnprotectedLabels    = map[string]string{common.GardenerDeletionProtected: "false"}
-			deletionConfirmedAnnotations = map[string]string{common.ConfirmationDeletion: "true"}
-		)
-
-		BeforeEach(func() {
-			c, err = client.New(restConfig, client.Options{Scheme: kubernetes.SeedScheme})
-			Expect(err).NotTo(HaveOccurred())
-
-			ns := &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}}
-			if err := c.Create(ctx, ns); err != nil && !apierrors.IsAlreadyExists(err) {
-				Expect(err).NotTo(HaveOccurred())
-			}
-
-			By("applying CRDs")
-			applier, err := kubernetes.NewChartApplierForConfig(restConfig)
-			Expect(err).NotTo(HaveOccurred())
-			repoRoot := filepath.Join("..", "..")
-			Expect(applier.Apply(ctx, filepath.Join(repoRoot, "charts", "seed-bootstrap", "charts", "extensions"), "extensions", "")).To(Succeed())
-
-			Eventually(func() bool {
-				for _, object := range extensionObjects {
-					err := c.Get(ctx, client.ObjectKeyFromObject(object), object)
-					if meta.IsNoMatchError(err) {
-						return false
-					}
-				}
-				return true
-			}, 1*time.Second, 50*time.Millisecond).Should(BeTrue())
-		})
-
-		objectID := func(obj client.Object) string {
-			return fmt.Sprintf("%T/%s", obj, client.ObjectKeyFromObject(obj))
-		}
-
-		testDeletionUnconfirmed := func(ctx context.Context, obj client.Object) {
-			Eventually(func() string {
-				err := c.Delete(ctx, obj)
-				return string(apierrors.ReasonForError(err))
-			}, 1*time.Second, 50*time.Millisecond).Should(ContainSubstring("annotation to delete"), objectID(obj))
-		}
-
-		testDeletionConfirmed := func(ctx context.Context, obj client.Object) {
-			Eventually(func() error {
-				return c.Delete(ctx, obj)
-			}, 1*time.Second, 50*time.Millisecond).ShouldNot(HaveOccurred(), objectID(obj))
-			Eventually(func() bool {
-				err := c.Get(ctx, client.ObjectKeyFromObject(obj), obj)
-				return apierrors.IsNotFound(err) || meta.IsNoMatchError(err)
-			}, 1*time.Second, 50*time.Millisecond).Should(BeTrue(), objectID(obj))
-		}
-
-		Context("custom resource definitions", func() {
-			It("should admit the deletion because CRD has no protection label", func() {
-				for _, obj := range crdObjects {
-					// patch out default gardener.cloud/deletion-protected=true label
-					_, err := controllerutil.CreateOrPatch(ctx, c, obj, func() error {
-						obj.SetLabels(nil)
-						return nil
-					})
-					Expect(err).NotTo(HaveOccurred(), objectID(obj))
-					testDeletionConfirmed(ctx, obj)
-				}
-			})
-
-			It("should admit the deletion because CRD's protection label is not true", func() {
-				for _, obj := range crdObjects {
-					// override default gardener.cloud/deletion-protected=true label
-					_, err := controllerutil.CreateOrPatch(ctx, c, obj, func() error {
-						obj.SetLabels(deletionUnprotectedLabels)
-						return nil
-					})
-					Expect(err).NotTo(HaveOccurred(), objectID(obj))
-					testDeletionConfirmed(ctx, obj)
-				}
-			})
-
-			It("should prevent the deletion because CRD's protection label is true but deletion is not confirmed", func() {
-				// CRDs in seed-bootstrap chart should have gardener.cloud/deletion-protected=true label by default
-				for _, obj := range crdObjects {
-					testDeletionUnconfirmed(ctx, obj)
-				}
-			})
-
-			It("should admit the deletion because CRD's protection label is true and deletion is confirmed", func() {
-				// CRDs in seed-bootstrap chart should have gardener.cloud/deletion-protected=true label by default
-				for _, obj := range crdObjects {
-					_, err := controllerutil.CreateOrPatch(ctx, c, obj, func() error {
-						obj.SetAnnotations(deletionConfirmedAnnotations)
-						return nil
-					})
-					Expect(err).NotTo(HaveOccurred(), objectID(obj))
-					testDeletionConfirmed(ctx, obj)
-				}
-			})
-		})
-
-		Context("extension resources", func() {
-			BeforeEach(func() {
-				By("creating extension test objects")
-				_, err := test.EnsureTestResources(ctx, c, "testdata")
-				Expect(err).NotTo(HaveOccurred())
-			})
-
-			It("should prevent the deletion because deletion is not confirmed", func() {
-				for _, obj := range extensionObjects {
-					testDeletionUnconfirmed(ctx, obj)
-				}
-			})
-
-			It("should admit the deletion because deletion is confirmed", func() {
-				for _, obj := range extensionObjects {
-					_, err := controllerutil.CreateOrPatch(ctx, c, obj, func() error {
-						obj.SetAnnotations(deletionConfirmedAnnotations)
-						return nil
-					})
-					Expect(err).NotTo(HaveOccurred(), objectID(obj))
-					testDeletionConfirmed(ctx, obj)
-				}
-			})
-		})
-
-		Context("other resources", func() {
-			It("should not block deletion of other resources", func() {
-				Expect(c.Create(ctx, podObject)).To(Succeed())
-				testDeletionConfirmed(ctx, podObject)
-			})
-		})
-	})
 })
+
+func expectAllowed(response admission.Response, reason gomegatypes.GomegaMatcher, optionalDescription ...interface{}) {
+	Expect(response.Allowed).To(BeTrue(), optionalDescription...)
+	Expect(string(response.Result.Reason)).To(reason, optionalDescription...)
+}
+
+func expectDenied(response admission.Response, reason gomegatypes.GomegaMatcher, optionalDescription ...interface{}) {
+	Expect(response.Allowed).To(BeFalse(), optionalDescription...)
+	Expect(response.Result.Code).To(BeEquivalentTo(http.StatusForbidden), optionalDescription...)
+	Expect(string(response.Result.Reason)).To(reason, optionalDescription...)
+}
+
+func expectErrored(response admission.Response, code, err gomegatypes.GomegaMatcher, optionalDescription ...interface{}) {
+	Expect(response.Allowed).To(BeFalse(), optionalDescription...)
+	Expect(response.Result.Code).To(code, optionalDescription...)
+	Expect(response.Result.Message).To(err, optionalDescription...)
+}
